Add libunbound/remote.h
Add a shared header containing prototypes for functions that both ends
of a remote control connection need to implement.
